LUCENE-8725: Make TermsQuery.SeekingTermSetTermsEnum a top level class and public (#21)
