+++ lucene/join/src/java/org/apache/lucene/search/join/TermsQuery.java
@@ -19,13 +19,11 @@ package org.apache.lucene.search.join;
 import java.io.IOException;
 import java.util.Objects;
 
-import org.apache.lucene.index.FilteredTermsEnum;
 import org.apache.lucene.index.Terms;
 import org.apache.lucene.index.TermsEnum;
 import org.apache.lucene.search.MultiTermQuery;
 import org.apache.lucene.search.Query;
 import org.apache.lucene.util.AttributeSource;
-import org.apache.lucene.util.BytesRef;
 import org.apache.lucene.util.BytesRefHash;
 
 /**
@@ -96,76 +94,6 @@ public class TermsQuery extends MultiTermQuery {
   @Override
   public int hashCode() {
     return classHash() + Objects.hash(field, fromField, fromQuery, indexReaderContextId);
-  }
-
-  static class SeekingTermSetTermsEnum extends FilteredTermsEnum {
-
-    private final BytesRefHash terms;
-    private final int[] ords;
-    private final int lastElement;
-
-    private final BytesRef lastTerm;
-    private final BytesRef spare = new BytesRef();
-
-    private BytesRef seekTerm;
-    private int upto = 0;
-
-    SeekingTermSetTermsEnum(TermsEnum tenum, BytesRefHash terms, int[] ords) {
-      super(tenum);
-      this.terms = terms;
-      this.ords = ords;
-      lastElement = terms.size() - 1;
-      lastTerm = terms.get(ords[lastElement], new BytesRef());
-      seekTerm = terms.get(ords[upto], spare);
-    }
-
-    @Override
-    protected BytesRef nextSeekTerm(BytesRef currentTerm) throws IOException {
-      BytesRef temp = seekTerm;
-      seekTerm = null;
-      return temp;
-    }
-
-    @Override
-    protected AcceptStatus accept(BytesRef term) throws IOException {
-      if (term.compareTo(lastTerm) > 0) {
-        return AcceptStatus.END;
-      }
-
-      BytesRef currentTerm = terms.get(ords[upto], spare);
-      if (term.compareTo(currentTerm) == 0) {
-        if (upto == lastElement) {
-          return AcceptStatus.YES;
-        } else {
-          seekTerm = terms.get(ords[++upto], spare);
-          return AcceptStatus.YES_AND_SEEK;
-        }
-      } else {
-        if (upto == lastElement) {
-          return AcceptStatus.NO;
-        } else { // Our current term doesn't match the the given term.
-          int cmp;
-          do { // We maybe are behind the given term by more than one step. Keep incrementing till we're the same or higher.
-            if (upto == lastElement) {
-              return AcceptStatus.NO;
-            }
-            // typically the terms dict is a superset of query's terms so it's unusual that we have to skip many of
-            // our terms so we don't do a binary search here
-            seekTerm = terms.get(ords[++upto], spare);
-          } while ((cmp = seekTerm.compareTo(term)) < 0);
-          if (cmp == 0) {
-            if (upto == lastElement) {
-              return AcceptStatus.YES;
-            }
-            seekTerm = terms.get(ords[++upto], spare);
-            return AcceptStatus.YES_AND_SEEK;
-          } else {
-            return AcceptStatus.NO_AND_SEEK;
-          }
-        }
-      }
-    }
-
   }
 
 }
