+++ solr/FULLSTORY-CHANGES.txt
@@ -4,3 +4,4 @@
 * FullStory: special OOMKILLER handling for spaces (Scott Blum)
 * Remove pathological error reporting. Don't serialize the entire clusterstate to log (Scott Blum)
 * Expose DocValuesTermsCollector and TermsQuery (Jaime Yap)
+* LUCENE-8725: Make TermsQuery.SeekingTermSetTermsEnum a top level class and public
